Move export-graph function into its own file
